Refactor: clique and votingmachine separate packages.
