+++ protocol/consensus/voter_test.go
@@ -7,9 +7,11 @@ import (
 	"github.com/relab/hotstuff/internal/proto/clientpb"
 	"github.com/relab/hotstuff/internal/testutil"
 	"github.com/relab/hotstuff/protocol"
+	"github.com/relab/hotstuff/protocol/clique"
 	"github.com/relab/hotstuff/protocol/consensus"
 	"github.com/relab/hotstuff/protocol/leaderrotation/fixedleader"
 	"github.com/relab/hotstuff/protocol/rules/chainedhotstuff"
+	"github.com/relab/hotstuff/protocol/votingmachine"
 	"github.com/relab/hotstuff/security/crypto/ecdsa"
 	"github.com/relab/hotstuff/wiring"
 )
@@ -48,7 +50,7 @@ func wireUpVoter(
 		consensusRules,
 	)
 	leaderRotation := fixedleader.New(2) // want a leader that is not 1
-	votingMachine := consensus.NewVotingMachine(
+	votingMachine := votingmachine.New(
 		depsCore.Logger(),
 		depsCore.EventLoop(),
 		depsCore.RuntimeCfg(),
@@ -56,7 +58,7 @@ func wireUpVoter(
 		depsSecurity.Authority(),
 		viewStates,
 	)
-	dissAgg := consensus.NewClique(
+	dissAgg := clique.New(
 		depsCore.RuntimeCfg(),
 		votingMachine,
 		leaderRotation,
